requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
filelock==3.17.0
fsspec==2025.3.0
huggingface-hub==0.29.2
idna==3.10
numpy==2.2.3
packaging==24.2
protobuf==6.30.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentencepiece==0.2.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
urllib3==2.3.0
